added color selection on metro page. created change-theme.js and moved function to it.
