+++ Metro/index.html
@@ -123,6 +123,13 @@
                 <select name="soundbank" id="select-soundbank">
                     <option value = "./data/soundboard.js">SoundBoard</option>
                 </select>
+                
+                <label for="color-scheme">Color Scheme:</label>
+                <select id="color-scheme" name="color-scheme">
+                    <option value="monochrome">Monochrome</option>
+                    <option value="halloween">Halloween</option>
+                    <option value="neon">Neon</option>
+                </select>
             </form>
 
             <form id="metronome-sound">Metronome Sound
